move deploymentProfilesReader to main process
